refactor(identity-server-domain): identity 서버 도메인 분리
