+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/identity/infrastructure/persistence/jpa/repository/CredentialAccountRepository.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.identity.infrastructure.persistence.jpa.repository;
 
-import com.seatliberator.seatliberator.identity.domain.CredentialAccount;
+import com.seatliberator.seatliberator.identity.server.domain.account.CredentialAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
