+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/jwks/application/config/JwksConfiguration.java
@@ -4,12 +4,12 @@ import com.nimbusds.jose.jwk.JWKSet;
 import com.nimbusds.jose.jwk.RSAKey;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.proc.SecurityContext;
+import com.seatliberator.seatliberator.identity.server.domain.jwks.RSASignatureKey;
 import com.seatliberator.seatliberator.jwks.application.port.in.KeyProvider;
 import com.seatliberator.seatliberator.jwks.application.port.in.TokenProvider;
 import com.seatliberator.seatliberator.jwks.application.port.out.ByteEncoder;
 import com.seatliberator.seatliberator.jwks.application.service.JwtProvider;
 import com.seatliberator.seatliberator.jwks.application.service.OpaqueTokenProvider;
-import com.seatliberator.seatliberator.jwks.domain.RSASignatureKey;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
