--- identity/identity-application/src/main/java/com/seatliberator/seatliberator/jwks/domain/KeyStatus.java
@@ -1,6 +1,0 @@
-package com.seatliberator.seatliberator.jwks.domain;
-
-public enum KeyStatus {
-    SIGNABLE,
-    VERIFY_ONLY
-}
