+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/role/application/service/UserGrantedRoleService.java
@@ -2,9 +2,9 @@ package com.seatliberator.seatliberator.role.application.service;
 
 import com.seatliberator.seatliberator.identity.core.role.NamespaceRole;
 import com.seatliberator.seatliberator.identity.core.role.NamespaceRoleFormatter;
+import com.seatliberator.seatliberator.identity.server.domain.role.UserGrantedRole;
 import com.seatliberator.seatliberator.role.application.port.in.*;
 import com.seatliberator.seatliberator.role.application.port.out.UserGrantedRoleStore;
-import com.seatliberator.seatliberator.role.domain.UserGrantedRole;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
