+++ identity/identity-application/src/main/java/com/seatliberator/seatliberator/role/infrastructure/persistence/jpa/JpaUserGrantedRoleStore.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.role.infrastructure.persistence.jpa;
 
+import com.seatliberator.seatliberator.identity.server.domain.role.UserGrantedRole;
 import com.seatliberator.seatliberator.role.application.port.out.UserGrantedRoleStore;
-import com.seatliberator.seatliberator.role.domain.UserGrantedRole;
 import com.seatliberator.seatliberator.role.infrastructure.persistence.jpa.repository.UserGrantedRoleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
