+++ identity/identity-server/build.gradle.kts
@@ -1,0 +1,6 @@
+plugins {
+    id("seatliberator.base.aggregate")
+}
+
+group = "com.seatliberator.seatliberator"
+version = "0.0.1-SNAPSHOT"
