+++ identity/identity-server/identity-server-domain/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("seatliberator.layer.domain")
+}
+
+group = "com.seatliberator.seatliberator"
+version = "0.0.1-SNAPSHOT"
+
+dependencies {
+    implementation(project(":identity:identity-core"))
+    implementation(project(":kernel:kernel-core"))
+
+    testImplementation(project(":kernel:kernel-test"))
+}
