+++ identity/identity-server/identity-server-domain/src/main/java/com/seatliberator/seatliberator/identity/server/domain/account/CredentialAccount.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.domain;
+package com.seatliberator.seatliberator.identity.server.domain.account;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
