+++ identity/identity-server/identity-server-domain/src/main/java/com/seatliberator/seatliberator/identity/server/domain/jwks/AbstractKeyMetadata.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.jwks.domain;
+package com.seatliberator.seatliberator.identity.server.domain.jwks;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
