+++ identity/identity-server/identity-server-domain/src/main/java/com/seatliberator/seatliberator/identity/server/domain/jwks/KeyStatus.java
@@ -1,0 +1,6 @@
+package com.seatliberator.seatliberator.identity.server.domain.jwks;
+
+public enum KeyStatus {
+    SIGNABLE,
+    VERIFY_ONLY
+}
